requirements.txt
asgiref==3.8.1
certifi==2023.7.22
cv==1.0.0
Django==5.1
django-environ==0.11.2
django-filter==24.3
django-rest-authtoken==2.1.4
djangorestframework==3.15.2
Markdown==3.7
numpy==1.26.2
opencv-python==4.8.1.78
psycopg2==2.9.9
psycopg2-binary==2.9.9
pygame==2.5.2
sqlparse==0.5.1
